Z4XV8R8 move the inline TUI commands out of commands.ts into commands/*.cmd.ts
